+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-
+### Added
+- Add `ceres ops webserver backup`
 
 
 ## [0.0.14] - 2018-05-28